requirements.txt
google-generativeai>=0.2.0
requests>=2.28.0
python-dotenv>=1.0.0
matplotlib>=3.6.0
numpy>=1.23.0
Pillow>=9.0.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3
tqdm>=4.65.0